refactor(scope-analyzer): extract declarations handlers to submodule
Moves NodeKind::VariableDeclaration, VariableListDeclaration, and Use
arms from analyze_node into declarations.rs.
